phase-2-web/backend/requirements.txt
fastapi>=0.110.0
uvicorn[standard]>=0.27.0
sqlmodel>=0.0.16
asyncpg>=0.29.0
python-dotenv>=1.0.0
pytest>=7.0.0
httpx>=0.27.0
pytest-asyncio>=0.23.0
